+++ app/__init__.py
@@ -1,6 +1,6 @@
 from flask import Flask
 from instance.config import app_config
-from app.api.v1.views import api as v1
+from app.api.v1.views.views import api as v1
 
 def create_app(config_name):
     """ Create flask application method """
